RavenDB-7592: * rearrange rvn (usage, commands : admin-channel + offline-operations, help) * move offline operations from cli back to rvn * move CliDelimiter from Sparrow to Raven.Client
